Rearraged somethings
